\name{barPlot3d}
\alias{barPlot3d}
\title{3 Dimensional Bar Ploting}
\description{
  A rgl-based function for 3 dimensinal bar ploting.
}
\arguments{
\item{mat}{ a numeric matrix represents the height and localtion of each
bar}
\item{topcol}{a color matrix with same dimensions of mat.}
\item{sidecol}{a color matrix with same dimensions of mat for side color
of bar.}
}
\usage{
barPlot3d(mat, topcol=1, sidecol = 1)
}
\value{NULL}
\author{Gang Chen}
\examples{
data(barPlot3d)
barPlot3d(m/60, color)
}
\keyword{barPlot3d}
